Bumped up version.
Now the script is better than its previous versions, i can say that its safer than before to use it as now neither it can be bypassed nor terminated.
i would say it still need many improvements like encryption and allowing the user to set their own passwd rather than the predefined ones.
I recommend copying and pasting the script code in your bash.bashrc that way it'll run on each startup.

Bugs:
can be bypassed as termux has failsafe mode, but the intruder cannot do much in that mode. im still looking for a way to fix it

To-Do:
*Encryption
*Flexiblity to set your own credentials
*Installion script so that user wont have to do the stuff himself
*fix for failsafe mode
*probably more cant remember :P