Bug fixes again
Ffs, i didnt even knew how many bugs living under my nose @_@
this bug caused the custom path not to be added to the cbpt var.
well, it was my fault i didn't gave enough importance to exit code -_-

it was like

if return code is not equal 0 then do the following.
but the function (cst_bpt)  was returning 0 ( 0 is true in this world ) so my while loop was not getting executed. *sigh*

anyway, another one killed.

we live in a society..
Fuck, riseupgamers