Backup_script: code improvements and removed useless functionality
Latterly, I was able to mess up my termux environment somehow and the backup script refused to restore an old backup.
Two bugs which I've discovered so far
Script work as expected from home directory only and, I know why. It'll be fixed in the next commit
Also,i am thinking of implemdntung some sort of emergency function in this script it will hopefully be able to work as i am expecting it to.
and, i will drop support of menu from the script in the future,it will solely rely on user supplied arguements.